fix(build): clean up compiler warnings under -Wall -Wextra
Address the remaining -Wall -Wextra diagnostics without changing runtime behaviour: cast signed/unsigned mismatches to the correct signedness in utils.c, logs.c and mmgr.c so buffer-length guards cannot be silently defeated by implicit unsigned promotion; add GCC-recognised /* fall through */ comments for intentional switch fallthroughs in readconfig.c and socketpuller.c; fix the misplaced 'static' qualifier in domainstatistic.c; and make the StableBuffer Add size expression consistently size_t.

+++ logs.c
@@ -45,7 +45,7 @@ int Log_Init(ConfigFileInfo *ConfigInfo, BOOL PrintScreen, BOOL Debug)
                  ConfigGetRawString(ConfigInfo, "LogFileFolder"),
                  PATH_SLASH_CH
                  )
-        >= sizeof(FilePath)
+        >= (int)sizeof(FilePath)
     )
     {
         return -36;
@@ -114,7 +114,7 @@ static void CheckLength(void)
                          FilePath,
                          CurrentNumber
                          )
-                >= sizeof(FileRenamed)
+                >= (int)sizeof(FileRenamed)
             )
             {
                 return;
